JFrame bug
WPILib py is very cool. Was playing around with the project and got hit
with a bug that we unearthed in our mods of robot builder. essentially
if you close RB while minimized your settings get written to the OS for
the window location to be -32000, -32000 which is a jFrame way of
saying you’re minimized resulting in your program to open off of your
screen.. figured i’d commit it in case it happens to any users of
wpilibpy